shifted the ingest.py ModCommand class into it's own module so I can re-use it
